refactor: renames files according to convention.  adds function documentation
