+++ src/js/containers/albums.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import Calendar from '../components/calendar';
 import { applyFilter, clearFilter } from '../actions/calendar.old';
-import filterFeed from '../lib/filterFeed';
+import filterFeed from '../lib/filter-feed';
 import getPastEvents from '../lib/getPastEvents';
 
 const mapStateToProps = (state) => {
