+++ src/js/containers/calendar.js
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux';
 import Calendar from '../components/calendar';
 import { applyFilter, clearFilter } from '../actions/calendar.old';
-import filterFeed from '../lib/filterFeed';
-import getFutureEvents from '../lib/getFutureEvents';
+import filterFeed from '../lib/filter-feed';
+import getFutureEvents from '../lib/get-future-events';
 
 
 const mapStateToProps = ({ calendar }) => {
