+++ src/js/containers/feed.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import Feed from '../components/feed';
 import { applyFilter, clearFilter } from '../actions/feed';
 import { getEvent } from '../actions/event/data';
-import filterFeed from '../lib/filterFeed';
+import filterFeed from '../lib/filter-feed';
 import Router from '../router';
 
 const mapStateToProps = ({ feed }) => {
