+++ src/js/lib/get-end-time.js
