--- src/js/lib/getUserID.js
@@ -1,4 +1,0 @@
-
-export default function getUserID () {
-  return '10156727442325251';
-}
